refactor: reorganize profile feature with clean architecture
- Create dedicated profile feature directory structure
- Break down profile screen into reusable widgets
- Move profile and edit screens to features/profile
- Create separate widget components:
  - ProfileHeader
  - ProfileDetails
  - ProfileMenu
- Improve code organization and maintainability
- Follow single responsibility principle
- Prepare structure for future testing
